Created Single service module
